Changes: added stretch to the pannable widget
RevBy: Vesa Halttunen
Details: in libmeegotouch 0.20.14 the pannable viewport stretches the widget it contains
to the viewport's size. Since the widget here needs to have empty space in the bottom,
a stretch is needed.

This patch also gets the size of a widget from the effective size hints and not from the geometry().