Fixes: NB#208728 - Sysuid process causing high wakeups when device is in idle state
RevBy: Vesa Halttunen
Details:
Status area scene was rendered even if display state is off while the renderer was created. Hence,
clock was still drawing itself even the state was off as it did not get any display changed calls.

However, once the displayState was changed, then the renderer was working as supposed to and clock
was updating depending on the display state. This patch now checks the initial display state during
renderer creation.